+++ src/TrafficManager/TrafficManager/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for the `RecordType` property on Traffic Manager profiles [#29711]
+    - Added `-RecordType` parameter to `New-AzTrafficManagerProfile` to allow specifying the DNS record type (`A`, `AAAA`, or `CNAME`) used for the profile's synthesized DNS records
+    - `RecordType` is now returned on the profile object from `Get-AzTrafficManagerProfile`, `New-AzTrafficManagerProfile`, and `Set-AzTrafficManagerProfile`
+    - Upgraded TrafficManager SDK to API version 2026-09-01
 
 ## Version 1.4.0
 * Added ChangeSafety Support